Add in a basic login system using redis
There are further security concerns here, but as the passwords
are randomly generated and essentially useless, this allows
the system to be developer and new VMs have a password without
worrying about it as much.

In the future, passwords will be hashed.
Longer term, VMs and accounts will be de-coupled.